fix/2678 push subscription manager (#2762)
* refactor(#2678): move PushSubscriptionManager to conductor-core

Move PushSubscriptionManager, PushSubscription, and PushPayload from
conductor-web/src/push.rs into conductor-core/src/push.rs so that
DB-touching manager logic lives in the library crate, not the binary.
Replace conductor-web/src/push.rs with a one-line re-export shim so
no callers in main.rs or routes/push.rs need to change.

Add web-push = "0.10" and tokio dev-dep to conductor-core; remove
web-push from conductor-web.


* refactor(#2678): alphabetize module declarations in conductor-core/src/lib.rs


* refactor(#2678): extract row_to_subscription helper to eliminate duplicated mapping closure


---------
